requirements.txt
Flask==2.3.3
Pillow==10.0.0
pytesseract==0.3.10
opencv-python-headless==4.8.0.74
numpy==1.24.3